Added Organize Method for Hearts
Not sure if it works entirely. We may need to come up with a system of delays so that we can animate the cards being swapped around if we wish. But for now this should organize cards by Diamonds, Clubs, Hearts, and Spades then their numbers within each of the suits.